alterações

+++ controllers/qrcode.controller.js
@@ -3,6 +3,8 @@ const dirQrcode = __dirname + "./../assets/images/qrcodes";
 const qrCode = require("qrcode");
 const fs = require("fs");
 
+const hubspotController = require("./hubspot.controller");
+
 function useQrcode(qrcode_id, company, callback) {
     const post = [new Date(), qrcode_id, company];
     connect.query("SELECT * FROM qrcode WHERE dataValidade > ? AND utilizacao > 0 AND idQRCode = ? AND empresa = ?", post, (err, rows) => {
@@ -12,16 +14,58 @@ function useQrcode(qrcode_id, company, callback) {
                 qrcode.utilizacao -= 1;
 
                 if (qrcode.utilizacao == 0) {
-                    connect.query(`DELETE FROM qrcode WHERE idQRCode = ${qrcode.idQRCode}`, (err, rows) => {
-                        if (!err) {
-                            const foto = `${dirQrcode}/${qrcode_id}.png`;
-                            fs.unlink(foto, (err) => {
-                                callback({
-                                    "statusCode": 200,
-                                    body: {
-                                        "message": "Valid QRCode"
-                                    }
-                                });
+                    hubspotController.getClient(qrcode.idUtilizador, (res) => {
+                        if (res.user) {
+                            const user = res.user;
+
+                            let properties = {};
+                            if (company == "Barquense") {
+                                if (qrcode.tipo_bilhete == "normal") {
+                                    properties.property = "bilhetes_disponiveis_barquense";
+                                    properties.value = parseInt(user.bilhetes_disponiveis_barquense) - 1;
+                                } else {
+                                    properties.property = "bilhetes_ida_e_volta_barquense";
+                                    properties.value = parseInt(user.bilhetes_ida_e_volta_barquense) - 1;
+                                }
+                            } else {
+                                if (qrcode.tipo_bilhete == "normal") {
+                                    properties.property = "bilhetes_disponiveis_transdev";
+                                    properties.value = parseInt(user.bilhetes_disponiveis_transdev) - 1;
+                                } else {
+                                    properties.property = "bilhetes_ida_e_volta_transdev";
+                                    properties.value = parseInt(user.bilhetes_ida_e_volta_transdev) - 1;
+                                }
+                            }
+                            hubspotController.updateClient(qrcode.idUtilizador, [properties], (res) => {
+                                if (res.statusCode == 200) {
+                                    connect.query(`DELETE FROM qrcode WHERE idQRCode = ${qrcode.idQRCode}`, (err, rows) => {
+                                        if (!err) {
+                                            const foto = `${dirQrcode}/${qrcode_id}.png`;
+                                            fs.unlink(foto, (err) => {
+                                                callback({
+                                                    "statusCode": 200,
+                                                    body: {
+                                                        "message": "Valid QRCode"
+                                                    }
+                                                });
+                                            })
+                                        } else {
+                                            callback({
+                                                "statusCode": 400,
+                                                body: {
+                                                    "message": "Invalid QRCode"
+                                                }
+                                            });
+                                        }
+                                    })
+                                } else {
+                                    callback({
+                                        "statusCode": 400,
+                                        body: {
+                                            "message": "Invalid QRCode"
+                                        }
+                                    });
+                                }
                             })
                         } else {
                             callback({
@@ -129,6 +173,11 @@ function generateQrcode(user_id, company, utilization, callback) {
         utilizacao: utilization
     }
 
+    if (utilization == 2) {
+        post.tipo_bilhete = "ida e volta";
+    } else {
+        post.tipo_bilhete = "normal";
+    }
     connect.query("INSERT INTO qrcode SET ?", post, (err, rows) => {
         if (!err) {
             const qrcode_id = rows.insertId;
